[refactor] :recycle: update imports after relocating mongoService (#313)
